pyproject.toml
[build-system]
requires = [
    "setuptools",
]
build-backend = "setuptools.build_meta"

[project]
name = "setec-tester-programs"
version = "2.11.0"
description = "SETEC product test programs."
authors = [
    {name = "Stephen Bell", email = "stephen.bell@setec.com.au"},
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "attrs",
    "jsonrpclib-pelix",
    "pydispatcher",
    "pyserial",
    "setec-isplpc==1.*",
    "setec-libtester==2.*",
    "setec-tester==2.*",
    "setec-updi==1.*",
    "setec-utility[erp]==1.*",
]

[tool.setuptools.packages.find]
include = ["programs*", "share*"]
